View Management: Clean up code
It's been a while since I've messed with how views are displayed, and
since that time I've learned a few new things on how to organize views.
The code for determining the state of a view was pretty confusing. While
it's still a little convoluted, I think I've managed to get everything
looking a little better. I'd like to move `<ViewStack>` out of
`view_container.js` soon, but I'll save that for another time.